added documentation, and other small modifications
