+++ srcs/check_filename.c
@@ -3,6 +3,12 @@
 #include <stddef.h>
 #include <string.h>
 
+/**
+ * @brief Checks if a file name or path corresponds to a .obj file
+ *
+ * @param filename : Path of the file to evaluate
+ * @return Return true or false depending on the result
+ */
 bool	is_obj_filename_valid(const char *filename)
 {
 	const char	*extension;
