+++ srcs/parse_element.c
@@ -16,8 +16,10 @@ bool	is_line_valid(char **line_data, bool (*check)(char *), int min_params,
 	while (line_data[i])
 	{
 		if (!check(line_data[i]))
-			return (error("wrong parameter type", line_data[i], line_nb),
-				false);
+		{
+			error("wrong parameter type", line_data[i], line_nb);
+			return (false);
+		}
 		i++;
 	}
 	i -= 1;
@@ -36,9 +38,9 @@ t_vec3	parse_vertex(char **line_data, int line_nb, bool *failed)
 		*failed = true;
 		return (ft_init_vec3(0));
 	}
-	new_vertex.x = ft_atod(line_data[1]);
-	new_vertex.y = ft_atod(line_data[2]);
-	new_vertex.z = ft_atod(line_data[3]);
+	new_vertex.x = atof(line_data[1]);
+	new_vertex.y = atof(line_data[2]);
+	new_vertex.z = atof(line_data[3]);
 	ft_free_strtab(line_data);
 	return (new_vertex);
 }
@@ -53,9 +55,9 @@ t_vec3	parse_normal(char **line_data, int line_nb, bool *failed)
 		*failed = true;
 		return (ft_init_vec3(0));
 	}
-	new_vertex.x = ft_atod(line_data[1]);
-	new_vertex.y = ft_atod(line_data[2]);
-	new_vertex.z = ft_atod(line_data[3]);
+	new_vertex.x = atof(line_data[1]);
+	new_vertex.y = atof(line_data[2]);
+	new_vertex.z = atof(line_data[3]);
 	ft_free_strtab(line_data);
 	return (new_vertex);
 }
@@ -70,8 +72,8 @@ t_vec2	parse_uv(char **line_data, int line_nb, bool *failed)
 		*failed = true;
 		return (ft_init_vec2(0));
 	}
-	new_vertex.x = ft_atod(line_data[1]);
-	new_vertex.y = ft_atod(line_data[2]);
+	new_vertex.x = atof(line_data[1]);
+	new_vertex.y = atof(line_data[2]);
 	ft_free_strtab(line_data);
 	return (new_vertex);
 }
